fix test with TRAVIS
Former-commit-id: e96722de67e05fdf3430fc5d19e4374b1e79c2a8 [formerly a4a267e5bee1462423c855a9c3a8518f8a39d6de [formerly f0a810b5221c8bae4b24d34a395f3968a10df40f]]
Former-commit-id: 59736a7ccb8675d277c32122021ece0dcfdbcde9
Former-commit-id: f85efca5b7a9b9f8dd85cda0dd3cfeb8476a0002

+++ connect/connect-jdbc-oracle12-sink/oracle12-sink.sh
@@ -10,23 +10,27 @@ then
      exit 1
 fi
 
-if test -z "$(docker images -q oracle/database:12.2.0.1-ee)"
+if [ -z "$TRAVIS" ]
 then
-     if [ ! -f ${DIR}/linuxx64_12201_database.zip ]
+     # not running with travis
+     if test -z "$(docker images -q oracle/database:12.2.0.1-ee)"
      then
-          logerror "ERROR: ${DIR}/linuxx64_12201_database.zip is missing. It must be downloaded manually in order to acknowledge user agreement"
-          exit 1
-     fi
-     log "Building oracle/database:12.2.0.1-ee docker image..it can take a while...(more than 15 minutes!)"
-     OLDDIR=$PWD
-     rm -rf ${DIR}/docker-images
-     git clone https://github.com/oracle/docker-images.git
+          if [ ! -f ${DIR}/linuxx64_12201_database.zip ]
+          then
+               logerror "ERROR: ${DIR}/linuxx64_12201_database.zip is missing. It must be downloaded manually in order to acknowledge user agreement"
+               exit 1
+          fi
+          log "Building oracle/database:12.2.0.1-ee docker image..it can take a while...(more than 15 minutes!)"
+          OLDDIR=$PWD
+          rm -rf ${DIR}/docker-images
+          git clone https://github.com/oracle/docker-images.git
 
-     cp ${DIR}/linuxx64_12201_database.zip ${DIR}/docker-images/OracleDatabase/SingleInstance/dockerfiles/12.2.0.1/linuxx64_12201_database.zip
-     cd ${DIR}/docker-images/OracleDatabase/SingleInstance/dockerfiles
-     ./buildDockerImage.sh -v 12.2.0.1 -e
-     rm -rf ${DIR}/docker-images
-     cd ${OLDDIR}
+          cp ${DIR}/linuxx64_12201_database.zip ${DIR}/docker-images/OracleDatabase/SingleInstance/dockerfiles/12.2.0.1/linuxx64_12201_database.zip
+          cd ${DIR}/docker-images/OracleDatabase/SingleInstance/dockerfiles
+          ./buildDockerImage.sh -v 12.2.0.1 -e
+          rm -rf ${DIR}/docker-images
+          cd ${OLDDIR}
+     fi
 fi
 
 export ORACLE_IMAGE="oracle/database:12.2.0.1-ee"
